re-org code some
